+++ core/state/intermediate_hashes.go
@@ -1,6 +1,10 @@
 package state
 
 import (
+	"bytes"
+	"fmt"
+
+	"github.com/ledgerwatch/bolt"
 	"github.com/ledgerwatch/turbo-geth/common"
 	"github.com/ledgerwatch/turbo-geth/common/dbutils"
 	"github.com/ledgerwatch/turbo-geth/common/pool"
@@ -28,6 +32,7 @@ func (ih *IntermediateHashes) WillUnloadBranchNode(prefixAsNibbles []byte, nodeH
 	}
 
 	key := pool.GetBuffer(keyBufferSize)
+	defer pool.PutBuffer(key)
 	trie.CompressNibbles(prefixAsNibbles, &key.B)
 
 	if err := ih.putter.Put(dbutils.IntermediateTrieHashBucket, common.CopyBytes(key.B), common.CopyBytes(nodeHash[:])); err != nil {
@@ -42,10 +47,101 @@ func (ih *IntermediateHashes) BranchNodeLoaded(prefixAsNibbles []byte) {
 	}
 
 	key := pool.GetBuffer(keyBufferSize)
+	defer pool.PutBuffer(key)
 	trie.CompressNibbles(prefixAsNibbles, &key.B)
 
-	if err := ih.deleter.Delete(dbutils.IntermediateTrieHashBucket, key.B); err != nil {
+	if err := ih.deleter.Delete(dbutils.IntermediateTrieHashBucket, common.CopyBytes(key.B)); err != nil {
 		log.Warn("could not delete intermediate trie hash", "err", err)
 		return
 	}
+}
+
+// PutTombstoneForDeletedAccount - placing tombstone only if given account has storage in database
+func PutTombstoneForDeletedAccount(db ethdb.MinDatabase, addrHash []byte) error {
+	if len(addrHash) != common.HashLength {
+		return nil
+	}
+
+	var boltDb *bolt.DB
+	if hasKV, ok := db.(ethdb.HasKV); ok {
+		boltDb = hasKV.KV()
+	} else {
+		return fmt.Errorf("only Bolt supported yet, given: %T", db)
+	}
+
+	hasIH := false
+	if err := boltDb.View(func(tx *bolt.Tx) error {
+		k, _ := tx.Bucket(dbutils.IntermediateTrieHashBucket).Cursor().Seek(addrHash)
+		if !bytes.HasPrefix(k, addrHash) {
+			k = nil
+		}
+		if k != nil {
+			hasIH = true
+		}
+
+		return nil
+	}); err != nil {
+		return err
+	}
+
+	if !hasIH {
+		return nil
+	}
+
+	return db.Put(dbutils.IntermediateTrieHashBucket, common.CopyBytes(addrHash), []byte{})
+}
+
+func ClearTombstonesForNewStorage(db ethdb.MinDatabase, storageKeyNoInc []byte) error {
+	var boltDb *bolt.DB
+	if hasKV, ok := db.(ethdb.HasKV); ok {
+		boltDb = hasKV.KV()
+	} else {
+		return fmt.Errorf("only Bolt supported yet, given: %T", db)
+	}
+
+	var toPut [][]byte
+	//var toDelete [][]byte
+
+	if err := boltDb.View(func(tx *bolt.Tx) error {
+		c := tx.Bucket(dbutils.IntermediateTrieHashBucket).Cursor()
+		for k, v := c.Seek(storageKeyNoInc[:common.HashLength]); k != nil; k, v = c.Next() {
+			if !bytes.HasPrefix(k, storageKeyNoInc[:common.HashLength]) {
+				k = nil
+			}
+			if k == nil {
+				break
+			}
+
+			isTombstone := v != nil && len(v) == 0
+			if isTombstone {
+				continue
+			}
+
+			for i := common.HashLength; i < len(k); i++ {
+				if storageKeyNoInc[i] == k[i] {
+					continue
+				}
+
+				toPut = append(toPut, common.CopyBytes(k[:i+1]))
+				break
+			}
+		}
+		return nil
+	}); err != nil {
+		return err
+	}
+
+	for _, k := range toPut {
+		if err := db.Put(dbutils.IntermediateTrieHashBucket, k, []byte{}); err != nil {
+			return err
+		}
+	}
+
+	for i := common.HashLength; i <= len(storageKeyNoInc); i++ {
+		if err := db.Delete(dbutils.IntermediateTrieHashBucket, common.CopyBytes(storageKeyNoInc[:i])); err != nil {
+			return err
+		}
+	}
+
+	return nil
 }
